backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
asyncpg>=0.30
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
httpx>=0.27
loguru>=0.7
python-multipart>=0.0.12
